refactors browse overlay logic into browseOverlayService
