+++ src/app/services/browse-map.service.ts
@@ -12,11 +12,9 @@ import { Layer, Vector } from 'ol/layer';
 import Polygon from 'ol/geom/Polygon';
 import { Extent, getCenter } from 'ol/extent';
 import VectorSource from 'ol/source/Vector';
-import WKT from 'ol/format/WKT';
 import LayerGroup from 'ol/layer/Group';
-import { Collection } from 'ol';
 import Projection from 'ol/proj/Projection';
-import { WktService } from '@services';
+import { BrowseOverlayService, WktService } from '@services';
 interface Dimension {
   width: number;
   height: number;
@@ -36,15 +34,14 @@ export class BrowseMapService {
   private view: View;
   private pinnedProducts: LayerGroup;
 
-  public constructor(private wktService: WktService) {}
+  public constructor(private wktService: WktService,
+    private browseOverlayService: BrowseOverlayService) {}
 
   public setMapBrowse(browse: string, wkt: string = ''): void {
-    const format = new WKT();
-    const feature = format.readFeature(wkt, {dataProjection: 'EPSG:4326',
-    featureProjection: 'EPSG:3857'});
+    const feature = this.wktService.wktToFeature(wkt, 'EPSG:3857');
     const polygon: Polygon = feature.getGeometry() as Polygon;
 
-    this.fixPolygonAntimeridian(polygon);
+    this.wktService.fixPolygonAntimeridian(feature, wkt);
 
     const polygonVectorSource = new VectorSource({
       features: [feature],
@@ -57,13 +54,14 @@ export class BrowseMapService {
 
     const center = getCenter( polygon.getExtent());
 
-    const Imagelayer = new ImageLayer({
-      source: new Static({
-        url: browse,
-        imageExtent: polygon.getExtent(),
-      }),
-      opacity: this.browseLayer?.getOpacity() ?? 1.0
-    });
+    const Imagelayer = this.browseOverlayService.createImageLayer(browse, wkt);
+    // const Imagelayer = new ImageLayer({
+    //   source: new Static({
+    //     url: browse,
+    //     imageExtent: polygon.getExtent(),
+    //   }),
+    //   opacity: this.browseLayer?.getOpacity() ?? 1.0
+    // });
 
     const mapSource = new XYZ({
       url : `https://api.maptiler.com/maps/hybrid/{z}/{x}/{y}.jpg?key=bFwkahiCrAA0526OlsHS`,
@@ -107,80 +105,8 @@ export class BrowseMapService {
     this.browseLayer.setOpacity(opacity);
   }
 
-  public createImageLayer(url: string, wkt: string, className: string = 'ol-layer') {
-    const format = new WKT();
-    const feature = format.readFeature(wkt, {dataProjection: 'EPSG:4326',
-    featureProjection: 'EPSG:3857'});
-
-    const polygon: Polygon = feature.getGeometry() as Polygon;
-    this.fixPolygonAntimeridian(polygon);
-
-    const imagelayer = new ImageLayer({
-      source: new Static({
-        url,
-        imageExtent: polygon.getExtent(),
-      }),
-      className,
-      zIndex: 0,
-      extent: polygon.getExtent(),
-      opacity: this.browseLayer?.getOpacity() ?? 1.0
-    });
-
-    return imagelayer;
-  }
-
-  public togglePinnedProduct() {
-    const pinnedLayerGroup = this.pinnedProducts;
-    const pinnedLayers = (pinnedLayerGroup.getLayers() as Collection<ImageLayer>).getArray();
-
-    const removed = pinnedLayers.find(l =>
-      (l.getSource() as Static).getUrl() === (this.browseLayer.getSource() as Static).getUrl()
-    );
-
-    if (!removed) {
-      pinnedLayerGroup.getLayers().push(this.browseLayer);
-    } else {
-      this.pinnedProducts.getLayers().remove(removed);
-    }
-  }
-
   public setPinnedProducts(pinnedProductStates: {[product_id in string]: PinnedProduct}) {
-    // Built in method Collection.clear() causes flickering when pinning new product,
-    // have to keep track of pinned products as work around
-
-    const pinnedProductIds = Object.keys(pinnedProductStates);
-
-    if (pinnedProductIds.length === 0) {
-      this.pinnedProducts?.getLayers().clear();
-    } else {
-      this.unpinProducts(pinnedProductIds);
-      this.pinProducts(pinnedProductIds, pinnedProductStates);
-    }
-  }
-
-  private unpinProducts(product_ids: string[]) {
-    this.pinnedProducts.getLayers().forEach(
-      l => {
-        if (product_ids.includes(l?.get('product_id'))) {
-          this.pinnedProducts.getLayers().remove(l);
-        }
-      }
-    );
-  }
-
-  private pinProducts(product_ids: string[], pinned: {[product_id in string]: PinnedProduct}) {
-    const imageLayers = product_ids.reduce((prev, product_id) => {
-      const current = prev;
-      const pinnedProd = this.createImageLayer(pinned[product_id].url, pinned[product_id].wkt, 'product_pin');
-      pinnedProd.set('product_id', product_id);
-      current.push(pinnedProd);
-      return current;
-    }, new Collection<ImageLayer>());
-
-    imageLayers.forEach( l => {
-        this.pinnedProducts.getLayers().push(l);
-      }
-    );
+    this.browseOverlayService.setPinnedProducts(pinnedProductStates, this.pinnedProducts);
   }
 
   public unpinAll() {
@@ -241,11 +167,5 @@ export class BrowseMapService {
     } else {
       this.map = this.newMap(view, [layer]);
     }
-  }
-
-  private fixPolygonAntimeridian(polygon: Polygon) {
-    polygon.setCoordinates(
-      [this.wktService.fixAntimeridianCoordinates(polygon.getCoordinates()[0])]
-      );
   }
 }
